feat: add Stores management feature with table and search functionality
- Implemented StoresTable component with status rendering and email fallback.
- Created StoreSearchBar for filtering stores.
- Developed StoresPage to manage store listings with loading and error states.
- Added useStores hook for fetching store data from the API.
- Introduced StoresService for API interactions related to stores.
- Created CSS styles for StoresTable and StoresPage.
- Added Storybook stories and tests for StoresTable and StoresPage components.
- Refactored routing to include new admin store management pages.
- Removed deprecated StoreLayout and related components.
