+++ src/features/admin/stores/components/StoresTable/index.ts
@@ -1,0 +1,1 @@
+export * from './StoresTable';
