--- src/features/store/components/StoreSearchBar/index.ts
@@ -1,1 +1,0 @@
-export * from './StoreSearchBar';
