--- src/features/store/components/StoresTable/StoresTable.css
@@ -1,16 +1,0 @@
-.storesTableStatusActive,
-.storesTableStatusInactive {
-    @apply inline-flex rounded px-2 py-1 text-xs font-semibold;
-}
-
-.storesTableStatusActive {
-    @apply bg-centra-success/10 text-centra-success;
-}
-
-.storesTableStatusInactive {
-    @apply bg-centra-error/10 text-centra-error;
-}
-
-.storesTablePendingAction {
-    @apply text-centra-text/40;
-}
