--- src/features/store/components/StoresTable/index.ts
@@ -1,1 +1,0 @@
-export * from './StoresTable';
