--- src/layouts/AdminLayout/AdminLayout.css
@@ -1,11 +1,0 @@
-.adminLayout {
-    @apply min-h-screen;
-}
-
-.adminMainLayout {
-    @apply min-w-0;
-}
-
-.adminMainContent {
-    @apply w-full p-6;
-}
