--- src/layouts/StoreLayout/StoreLayout.css
@@ -1,11 +1,0 @@
-.storeLayout {
-    @apply min-h-screen;
-}
-
-.storeMainLayout {
-    @apply min-w-0;
-}
-
-.storeMainContent {
-    @apply flex w-full items-center justify-center p-6 text-center;
-}
