+++ src/router/router.tsx
@@ -1,13 +1,22 @@
 import { createBrowserRouter, Navigate } from 'react-router-dom';
 import { ProtectedRoute } from './ProtectedRoute';
-import { AdminLayout } from '@/layouts/AdminLayout';
-import { StoreLayout } from '@/layouts/StoreLayout';
+import { AppLayout } from '@/layouts/AppLayout';
 
-//TODO create not found page
-// import { NotFoundPage } from '@/pages/NotFoundPage';
 import { RootRedirect } from './RootRedirect';
 import LoginPage from '@/pages/auth/LoginPage';
-import { StoresPage } from '@/pages/store/StoresPage';
+import { StoresPage } from '@/pages/admin/stores/StoresPage';
+
+const adminMenuItems = [
+    { key: '/admin', label: 'Dashboard' },
+    { key: '/admin/stores', label: 'Tiendas' },
+    { key: '/admin/plans', label: 'Planes' },
+];
+
+const storeMenuItems = [
+    { key: '/tienda/stock', label: 'Stock' },
+    { key: '/tienda/pos', label: 'POS' },
+    { key: '/tienda/orders', label: 'Pedidos' },
+];
 
 export const router = createBrowserRouter([
     // ── Raíz ─────────────────────────────────────────
@@ -17,7 +26,6 @@ export const router = createBrowserRouter([
     },
 
     // ── Auth ──────────────────────────────────────────
-    //TODO: Include Authlayout with login page.
     {
         path: 'login',
         element: <LoginPage />,
@@ -29,13 +37,11 @@ export const router = createBrowserRouter([
         element: <ProtectedRoute allowedRoles={['SUPER_ADMIN']} />,
         children: [
             {
-                element: <AdminLayout />,
+                element: <AppLayout title="Backoffice Admin" menuItems={adminMenuItems} />,
                 children: [
                     { index: true, element: <Navigate to="dashboard" replace /> },
                     { path: 'dashboard', element: <div>Admin</div> },
                     { path: 'stores', element: <StoresPage /> },
-                    //TODO: Include admin pages.
-                    //{ path: 'dashboard', element: <AdminDashboardPage /> },
                 ],
             },
         ],
@@ -45,42 +51,21 @@ export const router = createBrowserRouter([
     {
         path: 'tienda',
         children: [
-            // Rutes STORE_ADMIN
             {
                 element: <ProtectedRoute allowedRoles={['STORE_ADMIN']} />,
                 children: [
                     {
                         element: (
-                            <StoreLayout>
+                            <AppLayout title="Mi Tienda" menuItems={storeMenuItems}>
                                 <div>Store</div>
-                            </StoreLayout>
+                            </AppLayout>
                         ),
                         children: [
                             { index: true, element: <Navigate to="dashboard" replace /> },
-                            //TODO: Include store admin pages.
-                            //{ path: 'dashboard', element: <StoreDashboardPage /> },
                         ],
                     },
                 ],
             },
-
-            // TODO: Routes shared between STORE_ADMIN y others (e.g. STORE_LOGISTICS)
-            /* {
-                element: <ProtectedRoute allowedRoles={['STORE_ADMIN']} />,
-                children: [
-                    {
-                        element: <StoreLayout />,
-                        children: [],
-                    },
-                ],
-            }, */
         ],
     },
-
-    // ── 404 ───────────────────────────────────────────
-    //TODO: Create a NotFoundPage and include it here.
-    /* {
-        path: '*',
-        element: <NotFoundPage />,
-    }, */
 ]);
